stack it

+++ esphome/components/esphome/ota/ota_esphome.h
@@ -84,7 +84,7 @@ class ESPHomeOTAComponent : public ota::OTAComponent {
   uint8_t ota_features_{0};
 #ifdef USE_OTA_PASSWORD
   std::unique_ptr<uint8_t[]> auth_buf_;
-  size_t auth_buf_pos_{0};
+  uint8_t auth_buf_pos_{0};
   uint8_t auth_type_{0};  // Store auth type to know which hasher to use
 #endif                    // USE_OTA_PASSWORD
 };
